refactor: extract renderScriptHistory to history module
